Stabilize shared memory watcher test
