+++ src/heart/utilities/reactivex/__init__.py
@@ -1,0 +1,4 @@
+"""Reactivex helpers for shared stream handling."""
+
+from heart.utilities.reactivex.stream_sharing import \
+    share_stream as share_stream
